test(settings): share fake DOM harness
