+++ tests/frontend_settings_resolution_editors_smoke.mjs
@@ -1,5 +1,6 @@
 import assert from "node:assert/strict";
 import { readFileSync } from "node:fs";
+import { createFakeDocument, descendants } from "./frontend_support/fake_dom.mjs";
 
 function dataModule(relativePath, replacements = {}) {
   let source = readFileSync(new URL(relativePath, import.meta.url), "utf8");
@@ -17,94 +18,6 @@ const resolutionEditorsModule = await import(
 );
 
 assert.deepEqual(Object.keys(resolutionEditorsModule), ["createResolutionEditors"]);
-
-class FakeElement {
-  constructor(tagName) {
-    this.tagName = String(tagName).toUpperCase();
-    this.children = [];
-    this.parentElement = null;
-    this.style = { cssText: "" };
-    this.listeners = new Map();
-    this.attributes = new Map();
-    this.textContent = "";
-    this.title = "";
-    this.type = "";
-    this.inputMode = "";
-    this.value = "";
-    this.placeholder = "";
-    this.spellcheck = true;
-    this.selected = false;
-    this.focused = false;
-  }
-
-  append(...children) {
-    for (const child of children) {
-      child.parentElement = this;
-      this.children.push(child);
-      if (this.tagName === "SELECT" && child.tagName === "OPTION" && child.selected) {
-        this.value = child.value;
-      }
-    }
-  }
-
-  setAttribute(name, value) {
-    this.attributes.set(name, String(value));
-  }
-
-  getAttribute(name) {
-    return this.attributes.get(name) ?? null;
-  }
-
-  addEventListener(type, handler) {
-    const handlers = this.listeners.get(type) || [];
-    handlers.push(handler);
-    this.listeners.set(type, handlers);
-  }
-
-  emit(type, event = {}) {
-    const nextEvent = {
-      target: this,
-      defaultPrevented: false,
-      preventDefault() {
-        this.defaultPrevented = true;
-      },
-      ...event,
-    };
-    for (const handler of this.listeners.get(type) || []) {
-      handler(nextEvent);
-    }
-    return nextEvent;
-  }
-
-  focus() {
-    this.focused = true;
-  }
-
-  blur() {
-    this.focused = false;
-    this.emit("blur");
-  }
-}
-
-class FakeDocument {
-  constructor() {
-    this.createdElements = [];
-  }
-
-  createElement(tagName) {
-    const element = new FakeElement(tagName);
-    this.createdElements.push(element);
-    return element;
-  }
-}
-
-function descendants(root) {
-  const values = [];
-  for (const child of root.children) {
-    values.push(child, ...descendants(child));
-  }
-  return values;
-}
 
 function findAll(root, tagName) {
   return [root, ...descendants(root)].filter(
@@ -125,7 +38,7 @@ const TEXT = {
   naiaResolutionScaleTip: "Scale help",
 };
 
-const document = new FakeDocument();
+const document = createFakeDocument();
 const internalValues = new Map([
   ["naia.resolution_mode", "bucket_fit"],
   ["naia.resolution_scale", "1,75"],
@@ -239,7 +152,7 @@ assert.ok(
   "Every internal update must preserve the text setting type",
 );
 
-const fallbackDocument = new FakeDocument();
+const fallbackDocument = createFakeDocument();
 const fallbackEditors = resolutionEditorsModule.createResolutionEditors({
   document: fallbackDocument,
   text: (key) => TEXT[key] ?? key,
